+++ modules/admin/src/base/Controller.php
@@ -19,8 +19,7 @@ class Controller extends \luya\web\Controller
     public $layout = '@admin/views/layouts/main';
 
     /**
-     * @var bool When enabling `$disablePermissionCheck` all actions are not secured by access controller but
-     *           are do require an authtenticated user (logged in user).
+     * @var bool When enabling `$disablePermissionCheck` all actions are not secured by access controller but are do require an authtenticated user (logged in user).
      */
     public $disablePermissionCheck = false;
 
@@ -33,7 +32,6 @@ class Controller extends \luya\web\Controller
      * + By default not logged in users.
      *
      * @return array Rule-Definitions
-     *
      * @see yii\filters\AccessControl
      */
     public function getRules()
@@ -44,6 +42,11 @@ class Controller extends \luya\web\Controller
                 'actions' => [], // apply to all actions by default
                 'roles' => ['@'],
                 'matchCallback' => function ($rule, $action) {
+                
+                	if (!Yii::$app->adminuser->isGuest) {
+                		Yii::$app->luyaLanguage = Yii::$app->adminuser->identity->setting->get('luyadminlanguage', Yii::$app->luyaLanguage);
+                	}
+                
                     // see if a controller property has been defined to disabled the permission checks
                     if ($action->controller->disablePermissionCheck) {
                         return true;